i2c_bus: add discharge function
For peripheral reset or power cycling, it can be necessary to hold SDA
and SCL low for a certain amount of time. Add a discharge() function
that lambdas can call to put the bus into a drive-low state.

The current implementation is simplistic and requires setup() to be
called to terminate discharge make the bus usable again.